native: macos: move utils to separate file
